Move backend challenge files to project/backend
