--- backend/challenges/package.json
@@ -1,23 +1,0 @@
-{
-  "scripts": {
-    "dev": "nodemon src/app.js"
-  },
-  "type": "module",
-  "dependencies": {
-    "@prisma/client": "^6.10.1",
-    "cors": "^2.8.5",
-    "dotenv": "^17.0.0",
-    "express": "^5.1.0",
-    "express-validator": "^7.2.1",
-    "nodeman": "^1.1.2",
-    "nodemon": "^3.1.10",
-    "prisma": "^6.10.1"
-  },
-  "name": "challenges",
-  "version": "1.0.0",
-  "main": "index.js",
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": ""
-}
