+++ project/backend/package.json
@@ -1,17 +1,23 @@
 {
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
   "scripts": {
-    "dev": "node --watch src/app.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "nodemon src/app.js"
   },
+  "type": "module",
+  "dependencies": {
+    "@prisma/client": "^6.10.1",
+    "cors": "^2.8.5",
+    "dotenv": "^17.0.0",
+    "express": "^5.1.0",
+    "express-validator": "^7.2.1",
+    "nodeman": "^1.1.2",
+    "nodemon": "^3.1.10",
+    "prisma": "^6.10.1"
+  },
+  "name": "challenges",
+  "version": "1.0.0",
+  "main": "index.js",
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "module",
-  "dependencies": {
-    "express": "^5.1.0"
-  }
+  "description": ""
 }
